Setup demo integration with GitLab
GitLab's Code Quality reporting uses a specific subset[2] of the
Code Climate output[1]. Here we implement the sections of the full
Code Climate spec we have easy access to, which includes all of
the fields that GitLab requests.

A GitLab CI script has been added with unit test, coverage, code
quality, and Sonar integrations demonstrated. The use of the reuse
tool is disabled, as it was having unclear issues with believing
the pip cache was part of the project.

A few minor bugs have also been quashed:
 - Prevented adding sub-modules as explicit base packages.
 - Add the creating Tool to annotations in the results processor.
 - Fix handling of streams when no Reporter requested any.
 - Exit with status code 1 is the run is not classified as a success.

[1] https://github.com/codeclimate/platform/blob/master/spec/analyzers/SPEC.md#issues
[2] https://docs.gitlab.com/ee/ci/testing/code_quality.html#implement-a-custom-tool